+++ app/src/main/res/menu/git_clone.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:pwstore="http://schemas.android.com/apk/res-auto"
-    tools:context="com.zeapo.pwdstore.git.GitActivity" >
+    tools:context="com.zeapo.pwdstore.git.GitServerConfigActivity" >
     <item android:id="@+id/user_pref"
         android:title="@string/action_settings"
         android:orderInCategory="100"
